more changes done
added a developers page in the developers folder, design the page to
have all its appropriate features, linked the developer to his/her
page, reduced the number of templates and so much more… check for
yourself and let me know if you like it so that we can move on. thanks.